Add files via upload
ceaser cipher was purposely broken, and each exercise requires to debug the script and find where the script broke and adjust the codes to make it work again. 

where the script broke, would usually show where the code needed to be fixed. the look at the code flow and repair the flow to work again